# ApiSwagger/obj/Debug/net8.0/EndpointInfo/ApiSwagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ApiSwagger",
    "version": "1.0"
  },
  "paths": {
    "/api/Inventario/ObtenerTodos": {
      "get": {
        "tags": [
          "Inventario"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Inventario/ObtenerPorId/{id}": {
      "get": {
        "tags": [
          "Inventario"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Inventario"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventario"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventario"
                }
              }
            }
          }
        }
      }
    },
    "/api/Inventario/InsertarProducto": {
      "post": {
        "tags": [
          "Inventario"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inventario"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Inventario"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Inventario"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Inventario/EditarProducto/{id}": {
      "put": {
        "tags": [
          "Inventario"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inventario"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Inventario"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Inventario"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Inventario/EliminarProducto/{id}": {
      "delete": {
        "tags": [
          "Inventario"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Inventario": {
        "required": [
          "nombre",
          "precio",
          "stock"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "precio": {
            "minimum": 0.01,
            "type": "number",
            "format": "double"
          },
          "stock": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}